Add files via upload
Your manager has got a request to build a global dashboard with some key metrics, that can be one source of truth for all regions. Your manager has assigned this task to you to collate all the Region-specific sheets in one file, clean all unwanted data and map it with the other master tables and create an automated dashboard solution that can serve all.